www/react: Move pushIntoMapOfArrays() to buildbot-ui package
