+++ www/react-ui/src/main.ts
@@ -34,6 +34,7 @@ export * from "./contexts/Topbar";
 export * from "./stores/TimeStore";
 export * from "./stores/TopbarStore";
 
+export * from "./util/Collections";
 export * from "./util/FavIcon";
 export * from "./util/Moment";
 export * from "./util/React";
